backend/requirements.txt
flask
flask-cors
pyjwt
flask-jwt-extended
flask-sqlalchemy